WARNING: Fix an API inconsistency in dw_notebook_page_destroy/set() These two APIs incorrectly referenced the page ID as an int not long. This change may cause ABI problems with programs that use these on some platforms. Recompile your apps that use these functions for safety.
git-svn-id: https://svn.netlabs.org/repos/dwindows/trunk@2921 81767d24-ef19-dc11-ae90-00e081727c95

+++ android/dw.cpp
@@ -5591,7 +5591,7 @@ unsigned long API dw_notebook_page_new(HWND handle, ULONG flags, int front)
  *          handle: Handle to the notebook widget.
  *          pageid: ID of the page to be destroyed.
  */
-void API dw_notebook_page_destroy(HWND handle, unsigned int pageid)
+void API dw_notebook_page_destroy(HWND handle, unsigned long pageid)
 {
     JNIEnv *env;
 
@@ -5640,7 +5640,7 @@ unsigned long API dw_notebook_page_get(HWND handle)
  *          handle: Handle to the notebook widget.
  *          pageid: ID of the page to be made visible.
  */
-void API dw_notebook_page_set(HWND handle, unsigned int pageid)
+void API dw_notebook_page_set(HWND handle, unsigned long pageid)
 {
     JNIEnv *env;
