+++ ios/dw.m
@@ -9096,10 +9096,10 @@ DW_FUNCTION_RESTORE_PARAM3(handle, HWND, DW_UNUSED(flags), ULONG, front, int)
  *          handle: Handle to the notebook widget.
  *          pageid: ID of the page to be destroyed.
  */
-DW_FUNCTION_DEFINITION(dw_notebook_page_destroy, void, HWND handle, unsigned int pageid)
+DW_FUNCTION_DEFINITION(dw_notebook_page_destroy, void, HWND handle, unsigned long pageid)
 DW_FUNCTION_ADD_PARAM2(handle, pageid)
 DW_FUNCTION_NO_RETURN(dw_notebook_page_destroy)
-DW_FUNCTION_RESTORE_PARAM2(handle, HWND, pageid, unsigned int)
+DW_FUNCTION_RESTORE_PARAM2(handle, HWND, pageid, unsigned long)
 {
     DWNotebook *notebook = handle;
     DWNotebookPage *notepage = _dw_notepage_from_id(notebook, pageid);
@@ -9158,10 +9158,10 @@ DW_FUNCTION_RESTORE_PARAM1(handle, HWND)
  *          handle: Handle to the notebook widget.
  *          pageid: ID of the page to be made visible.
  */
-DW_FUNCTION_DEFINITION(dw_notebook_page_set, void, HWND handle, unsigned int pageid)
+DW_FUNCTION_DEFINITION(dw_notebook_page_set, void, HWND handle, unsigned long pageid)
 DW_FUNCTION_ADD_PARAM2(handle, pageid)
 DW_FUNCTION_NO_RETURN(dw_notebook_page_set)
-DW_FUNCTION_RESTORE_PARAM2(handle, HWND, pageid, unsigned int)
+DW_FUNCTION_RESTORE_PARAM2(handle, HWND, pageid, unsigned long)
 {
     DWNotebook *notebook = handle;
     DWNotebookPage *notepage = _dw_notepage_from_id(notebook, pageid);
